backend/pyproject.toml
[project]
name = "data-processing-pipeline"
version = "0.1.0"
description = "Data processing and embedding pipeline for external API data"
authors = [{name = "AI Assistant", email = "claude@example.com"}]
dependencies = [
    "requests>=2.31.0",
    "cohere>=4.0.0",
    "qdrant-client>=1.9.0",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">=3.8.1"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]